ai/requirements.txt
albumentations
fastapi
numpy
opencv-python-headless
pillow
torch
torchvision
uvicorn